CRUD de marcas y categorias listo

+++ src/redux/productos/productosThunk.ts
@@ -2,8 +2,8 @@ import axios, { AxiosResponse } from "axios";
 import { finishLoadingAplication, finishLoadingData, startLoadingAplication, startLoadingData } from "../aplication/aplicationSlice";
 import { AppDispatch, RootState } from "../store";
 import api from "../../api/config";
-import { createProducto, deleteProducto, getAllCategorias, getAllLogs, getAllMarcas, getAllProductos, getAllUnidadesMedida, updateProducto } from "./productosSlice";
-import { hideNotification, showNotificationError, showNotificationSuccess } from "../notification/notificationSlice";
+import { createCategoria, createMarca, createProducto, deleteCategoria, deleteMarca, deleteProducto, getAllCategorias, getAllLogs, getAllMarcas, getAllProductos, getAllUnidadesMedida, updateCategoria, updateMarca, updateProducto } from "./productosSlice";
+import { hideNotification, showNotificationError, showNotificationSuccess, showNotificationWarning } from "../notification/notificationSlice";
 
 interface UpdateProductoInterface {
     id: string;
@@ -14,6 +14,11 @@ interface UpdateProductoInterface {
     marcaId?: string;
     unidadMedidaId?: string;
 }
+interface UpdateProductoImagenInterface {
+    id: string;
+    imagenUrl: string;
+    imagen: File|undefined
+}
 interface CreateProductoInterface {
     codigo: string;
     nombre: string;
@@ -21,6 +26,21 @@ interface CreateProductoInterface {
     categoriaId: string;
     marcaId: string;
     unidadMedidaId: string;
+}
+
+interface CreateMarcaInterface{
+    nombre: string,
+    origen?: string
+}
+interface UpdateMarcaInterface extends CreateMarcaInterface{
+    id: string
+}
+interface CreateCategoriaInterface{
+    nombre: string,
+    detalle?: string
+}
+interface UpdateCategoriaInterface extends CreateCategoriaInterface{
+    id: string
 }
 
 export const getAllProductosAPI = () => {
@@ -41,6 +61,7 @@ export const getAllProductosAPI = () => {
 export const createProductoAPI = (data: CreateProductoInterface, file: File|undefined) => {
     return async(dispatch: AppDispatch, getState: () => RootState) => {
         const {id: sucursalId, userData} = getState().Sucursal;
+        if(!sucursalId) return;
         const newProducto = new FormData();
         newProducto.append('sucursalId', sucursalId);
         newProducto.append('codigo', data.codigo);
@@ -84,6 +105,7 @@ export const createProductoAPI = (data: CreateProductoInterface, file: File|unde
 export const deleteProductoAPI = (id: string) => {
     return async (dispatch: AppDispatch, getState: () => RootState) => {
         const {id: sucursalId, userData} = getState().Sucursal;
+        if(!sucursalId) return;
         try {
             dispatch(startLoadingData());
             const response: AxiosResponse = await api.delete('productos-ms/delete-producto', 
@@ -109,15 +131,12 @@ export const deleteProductoAPI = (id: string) => {
 }
 export const updateProductoAPI = (updateData: UpdateProductoInterface) => {
     return async(dispatch: AppDispatch, getState: () => RootState) => {
-        const {id, userData} = getState().Sucursal;
+        const {id:sucursalId, userData} = getState().Sucursal;
+        if(!sucursalId) return;
         try {
             dispatch(startLoadingData());
             const response: AxiosResponse = await api.post('productos-ms/update-producto', 
-                {sucursalId:id,...updateData},
-                {
-                    headers: {"X-User-Id": userData.id}
-                }
-            
+                {sucursalId,...updateData}, { headers: {"X-User-Id": userData.id} }
             );
             const {data, message} = response.data;
             dispatch(updateProducto(data));
@@ -132,7 +151,48 @@ export const updateProductoAPI = (updateData: UpdateProductoInterface) => {
                 setTimeout( () => dispatch(hideNotification()), 5000 );
             }else console.log(error);
         }
+    }
+}
+export const updateProductoImagenAPI = (data: UpdateProductoImagenInterface) => {
+    return async(dispatch: AppDispatch, getState: () => RootState) => {
+        if( data.imagen ){
+            const {id: sucursalId, userData} = getState().Sucursal;
+            if(!sucursalId) return;
 
+            const updatedProductoImagen = new FormData();
+            updatedProductoImagen.append('id', data.id);
+            updatedProductoImagen.append('sucursalId', sucursalId);
+            updatedProductoImagen.append('imagen', data.imagen);
+            updatedProductoImagen.append('imagenUrl', data.imagenUrl);
+        
+            try {
+                dispatch(startLoadingData());
+                const response: AxiosResponse = await api.post('productos-ms/update-producto-imagen', 
+                    updatedProductoImagen,
+                    {
+                        headers: {
+                            "Content-Type": "multipart/form-data",
+                            "X-User-Id": userData.id
+                        }
+                    }
+                );
+                const {data, message} = response.data;
+                dispatch(updateProducto(data));
+                dispatch(showNotificationSuccess({tittle: 'MODIFICACIÓN DE IMAGEN', description: message}));
+                setTimeout( () => dispatch(hideNotification()), 5000 );
+                dispatch(finishLoadingData());
+            } catch (error) {
+                if( axios.isAxiosError(error) && error.response ){
+                    const {data} = error.response;
+                    dispatch(showNotificationError({tittle: 'MODIFICACIÓN DE IMAGEN', description: data.message}));
+                    dispatch(finishLoadingData());
+                    setTimeout( () => dispatch(hideNotification()), 5000 );
+                }else console.log(error);
+            }
+        }else{
+            dispatch(showNotificationWarning({tittle: 'MODIFICACIÓN DE IMAGEN', description: 'La imagen original no fue modificada'}));
+            setTimeout( () => dispatch(hideNotification()), 5000 );
+        }
     }
 }
 export const getAllMarcasAPI = () => {
@@ -150,19 +210,77 @@ export const getAllMarcasAPI = () => {
         }
     }
 }
-export const createMarcaAPI = () => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
+export const createMarcaAPI = ( createMarcaData: CreateMarcaInterface ) => {
+    return async(dispatch: AppDispatch, getState: () => RootState) => {
+        const { id: sucursalId, userData } = getState().Sucursal;
+        if( !sucursalId ) return;
+        try {
+            dispatch(startLoadingData());
+            const response: AxiosResponse = await api.post('productos-ms/create-marca', 
+                {sucursalId, ...createMarcaData},{headers: {"X-User-Id": userData.id}}
+            );
+            const {data, message} = response.data;
+            dispatch(createMarca(data));
+            dispatch(showNotificationSuccess({tittle: 'CREACIÓN DE MARCA', description: message}));
+            setTimeout( () => dispatch(hideNotification()), 5000 );
+            dispatch(finishLoadingData());
+        } catch (error) {
+            if( axios.isAxiosError(error) && error.response ){
+                const {data} = error.response;
+                dispatch(showNotificationError({tittle: 'CREACIÓN DE MARCA', description: data.message}));
+                dispatch(finishLoadingData());
+                setTimeout( () => dispatch(hideNotification()), 5000 );
+            }else console.log(error);            
+        }
+    }
+}
+export const deleteMarcaAPI = ( id: string ) => {
+    return async(dispatch: AppDispatch, getState: () => RootState) => {
+        const { id: sucursalId, userData } = getState().Sucursal;
+        if( !sucursalId ) return;
+        try {
+            dispatch(startLoadingData());
+            const response: AxiosResponse = await api.delete('productos-ms/delete-marca', 
+                {headers: {"X-User-Id": userData.id}, params: {id, sucursalId}}
+            );
+            const {data, message} = response.data;
+            dispatch(deleteMarca(data.id));
+            dispatch(showNotificationSuccess({tittle: 'ELIMINACIÓN DE MARCA', description: message}));
+            setTimeout( () => dispatch(hideNotification()), 5000 );
+            dispatch(finishLoadingData());
+        } catch (error) {
+            if( axios.isAxiosError(error) && error.response ){
+                const {data} = error.response;
+                dispatch(showNotificationError({tittle: 'ELIMINACIÓN DE MARCA', description: data.message}));
+                dispatch(finishLoadingData());
+                setTimeout( () => dispatch(hideNotification()), 5000 );
+            }else console.log(error);
+        }
 
     }
 }
-export const deleteMarcaAPI = () => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
-
-    }
-}
-export const updateMarcaAPI = () => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
-
+export const updateMarcaAPI = (updateMarcaData: UpdateMarcaInterface) => {
+    return async(dispatch: AppDispatch, getState: () => RootState) => {
+        const { id: sucursalId, userData } = getState().Sucursal;
+        if(!sucursalId) return;
+        try {
+            dispatch(startLoadingData());
+            const response: AxiosResponse = await api.post('productos-ms/update-marca', 
+                {sucursalId,...updateMarcaData}, { headers: {"X-User-Id": userData.id} }
+            );
+            const {data, message} = response.data;
+            dispatch(updateMarca(data));
+            dispatch(showNotificationSuccess({tittle: 'ACTUALIZACIÓN DE MARCA', description: message}));
+            setTimeout( () => dispatch(hideNotification()), 5000 );
+            dispatch(finishLoadingData());
+        } catch (error) {
+            if( axios.isAxiosError(error) && error.response ){
+                const {data} = error.response;
+                dispatch(showNotificationError({tittle: 'ACTUALIZACIÓN DE MARCA', description: data.message}));
+                dispatch(finishLoadingData());
+                setTimeout( () => dispatch(hideNotification()), 5000 );
+            }else console.log(error);
+        }
     }
 }
 export const getAllCategoriasAPI = () => {
@@ -180,19 +298,76 @@ export const getAllCategoriasAPI = () => {
         }
     }
 }
-export const createCategoriaAPI = () => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
-
+export const createCategoriaAPI = (createCategoriaData: CreateCategoriaInterface) => {
+    return async(dispatch: AppDispatch, getState: () => RootState) => {
+        const { id: sucursalId, userData } = getState().Sucursal;
+        if( !sucursalId ) return;
+        try {
+            dispatch(startLoadingData());
+            const response: AxiosResponse = await api.post('productos-ms/create-categoria', 
+                {sucursalId, ...createCategoriaData},{headers: {"X-User-Id": userData.id}}
+            );
+            const {data, message} = response.data;
+            dispatch(createCategoria(data));
+            dispatch(showNotificationSuccess({tittle: 'CREACIÓN DE CATEGORÍA', description: message}));
+            setTimeout( () => dispatch(hideNotification()), 5000 );
+            dispatch(finishLoadingData());
+        } catch (error) {
+            if( axios.isAxiosError(error) && error.response ){
+                const {data} = error.response;
+                dispatch(showNotificationError({tittle: 'CREACIÓN DE CATEGORÍA', description: data.message}));
+                dispatch(finishLoadingData());
+                setTimeout( () => dispatch(hideNotification()), 5000 );
+            }else console.log(error);            
+        }
     }
 }
-export const deleteCategoriaAPI = () => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
-
+export const deleteCategoriaAPI = ( id: string ) => {
+    return async(dispatch: AppDispatch, getState: () => RootState) => {
+        const { id: sucursalId, userData } = getState().Sucursal;
+        if( !sucursalId ) return;
+        try {
+            dispatch(startLoadingData());
+            const response: AxiosResponse = await api.delete('productos-ms/delete-categoria', 
+                {headers: {"X-User-Id": userData.id}, params: {id, sucursalId}}
+            );
+            const {data, message} = response.data;
+            dispatch(deleteCategoria(data.id));
+            dispatch(showNotificationSuccess({tittle: 'ELIMINACIÓN DE CATEGORÍA', description: message}));
+            setTimeout( () => dispatch(hideNotification()), 5000 );
+            dispatch(finishLoadingData());
+        } catch (error) {
+            if( axios.isAxiosError(error) && error.response ){
+                const {data} = error.response;
+                dispatch(showNotificationError({tittle: 'ELIMINACIÓN DE CATEGORÍA', description: data.message}));
+                dispatch(finishLoadingData());
+                setTimeout( () => dispatch(hideNotification()), 5000 );
+            }else console.log(error);
+        }
     }
 }
-export const updateCategoriaAPI = () => {
-    return (dispatch: AppDispatch, getState: () => RootState) => {
-
+export const updateCategoriaAPI = ( updateCategoriaData: UpdateCategoriaInterface ) => {
+    return async(dispatch: AppDispatch, getState: () => RootState) => {
+        const { id: sucursalId, userData } = getState().Sucursal;
+        if(!sucursalId) return;
+        try {
+            dispatch(startLoadingData());
+            const response: AxiosResponse = await api.post('productos-ms/update-categoria', 
+                {sucursalId,...updateCategoriaData}, { headers: {"X-User-Id": userData.id} }
+            );
+            const {data, message} = response.data;
+            dispatch(updateCategoria(data));
+            dispatch(showNotificationSuccess({tittle: 'ACTUALIZACIÓN DE CATEGORÍA', description: message}));
+            setTimeout( () => dispatch(hideNotification()), 5000 );
+            dispatch(finishLoadingData());
+        } catch (error) {
+            if( axios.isAxiosError(error) && error.response ){
+                const {data} = error.response;
+                dispatch(showNotificationError({tittle: 'ACTUALIZACIÓN DE CATEGORÍA', description: data.message}));
+                dispatch(finishLoadingData());
+                setTimeout( () => dispatch(hideNotification()), 5000 );
+            }else console.log(error);
+        }
     }
 }
 export const getAllUnidadAPI = () => {
